external_libs/orca-backend-vibe/pyproject.toml
[tool.poetry]
name = "orca-backend-vibe"
version = "0.1.0"
description = ""
authors = ["Amer Jod <amer.j.trading@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-multipart = "^0.0.20"
pydantic = {extras = ["email"], version = "^2.12.3"}
pandas = "^2.3.1"
loguru = "^0.7.3"
pathos = "^0.3.4"
toml = "^0.10.2"
python-dotenv = "^1.1.1"
requests = "^2.32.4"
redis = "^6.4.0"
black = "^25.1.0"
websocket-client = "^1.8.0"
beautifulsoup4 = "^4.13.5"
supabase = "^2.21.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.7.1"
python-dateutil = "^2.9.0"
bcrypt = "3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
